+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## In progress
 
+### Fixed
+
+- Fixed default format.ps1xml file which was filed with empty type.ps1xml content.
+
 ### Changed
 
 - Changed script structure, deleting function to avoid declaring parameters twice.